backend/requirements.txt
django==3.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
sqlparse==0.2.4
django-cors-headers==3.14.0
django-environ==0.11.2
pymongo==3.12.3
pytz
gunicorn==21.2.0
whitenoise==6.6.0
